% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_hello_world.R
\docType{data}
\name{fhcrcData}
\alias{fhcrcData}
\title{Old data used in the prostata model}
\format{
An object of class \code{list} of length 10.
}
\usage{
fhcrcData
}
\description{
Old data used in the prostata model
}
\keyword{datasets}
