% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_hello_world.R
\name{discountedInterval}
\alias{discountedInterval}
\title{Integrate a discounted value}
\usage{
discountedInterval(y, start, finish, dr)
}
\arguments{
\item{y}{the undiscounted value}

\item{start}{the start time}

\item{finish}{the finish time}

\item{dr}{discount rate, expressed as a percentage}
}
\description{
Integrate a discounted value
}
