% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibSim.R, R/rcpp_hello_world.R
\name{callCalibrationPerson}
\alias{callCalibrationPerson}
\alias{callPersonSimulation}
\alias{callSimplePerson}
\alias{callSimplePerson2}
\alias{callIllnessDeath}
\title{call CalibrationPerson example}
\usage{
callCalibrationPerson(
  seed = 12345,
  n = 500,
  runpar = c(4, 0.5, 0.05, 10, 3, 0.5),
  mc.cores = 1
)

callPersonSimulation(n = 20, seed = rep(12345, 6))

callSimplePerson(n = 10)

callSimplePerson2(n = 10)

callIllnessDeath(n = 10L, cure = 0.1, zsd = 0)
}
\arguments{
\item{seed}{random number seed}

\item{n}{number of simulations (default=10)}

\item{runpar}{parameters}

\item{mc.cores}{number of cores}

\item{cure}{probability of cure}

\item{zsd}{frailty standard deviation}
}
\value{
data-frame

data-frame

data-frame

data-frame

data-frame
}
\description{
Example that uses the RngStream random number generator

Example that uses the Mersenne-Twister random number generator

Example that uses the Mersenne-Twister random number generator

Example that uses the Mersenne-Twister random number generator
}
