% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqueue.R, R/rcpp_hello_world.R
\name{pqueue__new}
\alias{pqueue__new}
\alias{pqueue__push}
\alias{pqueue__pop}
\alias{pqueue__cancel}
\alias{pqueue__empty}
\alias{pqueue__clear}
\alias{callCalibrationSimulation}
\alias{r_create_current_stream}
\alias{r_remove_current_stream}
\alias{r_set_user_random_seed}
\alias{r_rng_advance_substream}
\alias{r_next_rng_substream}
\alias{r_get_user_random_seed}
\title{C++ function}
\value{
data-frame

No return value, called for side effects

No return value, called for side effects

No return value, called for side effects

No return value, called for side effects

No return value, called for side effects

No return value, called for side effects
}
\description{
C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function

C++ function
}
