% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_hello_world.R
\docType{package}
\name{microsimulation-package}
\alias{microsimulation-package}
\alias{microsimulation}
\title{microsimulation}
\description{
The packages provides implementations of discrete event simulations in both R
and C++.
}
\section{Introduction}{


The packages provides implementations of discrete event simulations in both R
and C++.
}

\references{
\url{https://github.com/mclements/microsimulation}
}
\seealso{
\code{\link{Rcpp}}
}
\author{
Mark Clements \email{mark.clements@ki.se}
}
