% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcpp_hello_world.R
\name{dMVK}
\alias{dMVK}
\alias{pMVK}
\alias{qMVK}
\alias{rMVK}
\alias{rMVK2}
\title{Probability density function for an MVK distribution}
\usage{
dMVK(x, A, B, delta)

pMVK(x, A, B, delta, lower.tail = TRUE)

qMVK(p, A, B, delta, lower.tail = TRUE, upper = 10000)

rMVK(n, A, B, delta, upper = 10000)

rMVK2(n, A, B, delta, x0 = 100)
}
\arguments{
\item{x}{double for the x value}

\item{A}{double for the A parameter}

\item{B}{double for the B parameter}

\item{delta}{double for the delta parameter}

\item{lower.tail}{logical for whether to show the lower tail of the distribution}

\item{p}{double for the x value}

\item{upper}{double for the upper range to be searched}

\item{n}{integer for the number of random draws}

\item{x0}{double initial search point}
}
\value{
double for the pdf

double for the CDF or survival

double for the CDF or survival

double as a random draw

double as a random draw
}
\description{
This uses Newton-Raphson type solver, which depends on a reasonable
initial value. Not exported.
}
